+++ design-pattern-assignment1/abstract-factory-ui-component-library/src/Factories/mac-os-ui.factory.ts
@@ -1,7 +1,7 @@
 import { MacOSButton } from "../Components/MacOS/mac-os-button";
 import { MacOSCheckbox } from "../Components/MacOS/mac-os-checkbox";
 import { MacOSTextField } from "../Components/MacOS/mac-os-text-field";
-import { UIFactory } from "../Interfaces/abtract-ui-factory.interface";
+import { UIFactory } from "../Interfaces/ui-factory.interface";
 import { UIComponent } from "../Interfaces/ui-component.interface";
 import { errorMessages } from "../Utils/Constants/error-messages";
 import { UIComponentType } from "../Utils/Enums/ui-component.enum";
