+++ design-pattern-assignment1/abstract-factory-ui-component-library/src/Interfaces/ui-factory.interface.ts
